requirements.txt
psycopg2-binary==2.8.6
dj-database-url==0.5.0
Django==3.0.3
whitenoise==5.2.0
pytz==2021.1
gunicorn==20.0.4
django-crispy-forms==1.11.1
django-shopping-cart==0.1
pillow==8.1.2
django-heroku==0.3.1
djangorestframework==3.12.2
django-filter==2.4.0
django-cors-headers==3.7.0